refactor: pages/Onboarding -> components로 이전
