--- apps/web/app/onboarding/page.tsx
@@ -1,3 +1,0 @@
-import Page from "@pages/Onboarding/Page";
-
-export default Page;
