+++ apps/web/app/page.tsx
@@ -1,3 +1,3 @@
-import Page from "@pages/Onboarding/Page";
+import Page from "@components/leader-onboarding/Page";
 
 export default Page;
