+++ apps/web/src/components/leader-onboarding/Page.css.ts
