+++ apps/web/src/components/leader-onboarding/atoms/SliderCard.css.ts
