+++ apps/web/src/components/leader-onboarding/molecules/Slider.css.ts
